pyproject.toml
[tool.poetry]
name = "Modad"
version = "0.1.0"
description = ""
authors = ["Jessie Liauw A Fong"]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.1"
gitpython = "^2.1"
dataclasses = "^0.6.0"
click = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^3.6"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
m2r = "^0.2.1"
